requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
configparser==5.3.0
Django==3.2.19
django-bootstrap-form==3.4
django-filter==23.2
django-jalali-date==1.0.1
django-online-users==0.3
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
jdatetime==4.1.1
ldap3==2.9.1
Markdown==3.4.3
mysqlclient==2.2.0
pyasn1==0.5.0
PyMySQL==1.1.0
pytz==2023.3
requests==2.31.0
simplejson==3.17.6
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
